requirements.txt
rich
request
argparse
bs4

